Refactor: Organise Java files into packages
