--- src/main/java/ToDo.java
@@ -1,5 +1,0 @@
-public class ToDo extends Task {
-    protected ToDo(String description, boolean isDone) {
-        super(description, isDone, TaskType.TODO);
-    }
-}
