+++ src/main/java/timitomo/enums/CommandType.java
@@ -1,0 +1,15 @@
+package timitomo.enums;
+
+import timitomo.exceptions.TimitomoException;
+
+public enum CommandType {
+    BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE;
+
+    public static CommandType getAction(String type) throws TimitomoException {
+        try {
+            return CommandType.valueOf(type.toUpperCase());
+        } catch (IllegalArgumentException e) {
+            throw new TimitomoException("I don't know what that means");
+        }
+    }
+}
