+++ src/main/java/timitomo/exceptions/TimitomoException.java
@@ -1,3 +1,5 @@
+package timitomo.exceptions;
+
 public class TimitomoException extends Exception {
     public TimitomoException(String message) {
         super(message);
