+++ src/main/java/timitomo/storage/Storage.java
@@ -1,4 +1,13 @@
-import java.io.*;
+package timitomo.storage;
+
+import timitomo.exceptions.TimitomoException;
+import timitomo.tasks.Task;
+
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileReader;
+import java.io.FileWriter;
+import java.io.IOException;
 import java.util.ArrayList;
 
 public class Storage {
