+++ src/main/java/timitomo/tasks/ToDo.java
@@ -1,0 +1,9 @@
+package timitomo.tasks;
+
+import timitomo.enums.TaskType;
+
+public class ToDo extends Task {
+    public ToDo(String description, boolean isDone) {
+        super(description, isDone, TaskType.TODO);
+    }
+}
